+++ lib/analysis/ai/interpreter.rs
@@ -13,6 +13,16 @@ pub struct Interpreter<D: domain::Domain<M, V>, M: domain::Memory<V>, V: domain:
     pub domain: D
 }
 
+impl<D: domain::Domain<M, V>, M: domain::Memory<V>, V: domain::Value> Interpreter<D, M, V> {
+    pub fn new(domain: D) -> Interpreter<D, M, V> {
+        Interpreter {
+            m: ::std::marker::PhantomData,
+            v: ::std::marker::PhantomData,
+            domain: domain
+        }
+    }
+}
+
 
 impl<'a, D, M, V> fixed_point::FixedPointAnalysis<'a, domain::State<M, V>> for Interpreter<D, M, V>
     where D: 'a + domain::Domain<M, V>,
@@ -69,4 +79,67 @@ impl<'a, D, M, V> fixed_point::FixedPointAnalysis<'a, domain::State<M, V>> for I
             
         state0.join(state1)
     }
+}
+
+
+#[test]
+fn test() {
+    fn run() -> Result<()> {
+        use analysis::ai::test_lattice::*;
+
+        let mut cfg = il::ControlFlowGraph::new();
+
+        let head_index = {
+            cfg.new_block()?.index()
+        };
+
+        let left_index = {
+            cfg.new_block()?.index()
+        };
+
+        let right_index = {
+            cfg.new_block()?.index()
+        };
+
+        let tail_index = {
+            let block = cfg.new_block()?;
+
+            block.assign(il::scalar("$a0", 32), il::expr_const(0x570000, 32));
+            block.assign(il::scalar("$a0", 32),
+                il::Expression::add(il::expr_scalar("$a0", 32).into(),
+                                    il::expr_const(0x7038, 32))?);
+
+            block.index()
+        };
+
+        cfg.set_entry(head_index)?;
+
+        cfg.unconditional_edge(head_index, left_index)?;
+        cfg.unconditional_edge(head_index, right_index)?;
+        cfg.unconditional_edge(left_index, tail_index)?;
+        cfg.unconditional_edge(right_index, tail_index)?;
+
+        let function = il::Function::new(0, cfg);
+
+        let domain = TestLatticeDomain {};
+
+        let interpreter = Interpreter::new(domain);
+
+        let results = fixed_point::fixed_point_forward(interpreter, &function)?;
+
+        let block = function.block(tail_index).unwrap();
+        let instruction = block.instruction(1).unwrap();
+        let rfl = il::RefFunctionLocation::Instruction(block, instruction);
+        let rpl = il::RefProgramLocation::new(&function, rfl);
+
+        let a0 = results[&rpl].variable(&il::scalar("$a0", 32)).unwrap();
+
+        println!("{:?}", a0);
+
+        assert_eq!(*a0, TestLattice::Constant(il::const_(0x577038, 32)));
+
+        Ok(())
+    }
+
+    run().unwrap();
 }