+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Changed BTree containers in analysis/use_def to Hash containers.
 * Moved store/load methods from `domain::Memory` to `domain::Domain`
 * Remove `endian()` method from `Translator` trait.
+* Removed required methods for `ai::memory::Memory`, since they overlap with
+`ai::domain::Value`.
 
 ## Fixes
 
@@ -20,3 +22,6 @@
 * Add an `into_` method for `domain::Expression` to help translate expressions between domains.
 * Add `symbolize`, `remove_variable`, and `memory_mut` functions to `domain::State`
 * Add support for `Mipsel`, as it's only a minor change from Mips.
+* Impl `analysis::ai::Interpreter` and add simple `new` method to create a new
+abstract interpreter.
+* Lots of tests.